requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
click==7.1.2
flake8==3.9.0
iniconfig==1.1.1
isort==5.8.0
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
protobuf==3.15.7
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.3
regex==2021.4.4
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
